requirements.txt
numpy==2.3.5
pillow==12.0.0
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
scipy==1.16.3
tkinterdnd2==0.4.3
tqdm==4.67.2
